Refactor playback fragments
